fixes on the generation of R processes

+++ compss/runtime/adaptors/nio/scripts/persistent_worker_clean.sh
@@ -11,18 +11,20 @@
 #
 
 kill_recursive() {
-  # Active processes
-  to_kill=$(ps --ppid "$1" -o pid= | awk '{ print $1 }' | tr '\n' ' ')
+  if [ $1 -gt 0 ]; then 
+     # Active processes
+     to_kill=$(ps --ppid "$1" -o pid= | awk '{ print $1 }' | tr '\n' ' ')
  
-  # Kill each child process of the parent
-  for fn in $to_kill; do
-    kill_recursive "$fn"
-  done
+     # Kill each child process of the parent
+     for fn in $to_kill; do
+        kill_recursive "$fn"
+     done
 
-  # Kill the parent process
-  if [ -z "$to_kill" ]; then
-    # echo "Killing $1"
-    kill -15 "$1"
+     # Kill the parent process
+     if [ -z "$to_kill" ]; then
+        echo "Killing $1"
+        kill -15 "$1"
+     fi
   fi
 }
